+++ Rawdata_Porfolio_2/Entity Framework/Domain/episode.cs
@@ -11,7 +11,7 @@ namespace Rawdata_Porfolio_2.Pages.Entity_Framework.Domain
         [Key]
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Parent_Id { get; set; }
+        public Int64 Parent_Id { get; set; }
         public Title Title { get; set; }
         public int Season { get; set; }
         public int Ep_Number { get; set; }
