+++ Rawdata_Porfolio_2/Entity Framework/Domain/ratings.cs
@@ -11,7 +11,7 @@ namespace Rawdata_Porfolio_2.Pages.Entity_Framework.Domain
         [Key]
         public int User_Id { get; set; }
         public User User { get; set; }
-        public int Title_Id { get; set; }
+        public Int64 Title_Id { get; set; }
         public Title Title { get; set; }
         public int RatingOfTitle { get; set; }
         public DateTime Timestamp { get; set; }
